\name{runLift.hg19to38}
\alias{runLift.hg19to38}
\alias{RUNLIFT.HG19TO38}
\title{
    A short description (one line)
}
\description{
    A longer description
}
\usage{
    runLift.hg19to38(hg19.bed.filenames)
}
\arguments{
    \item{hg19.bed.filenames}{
        one or more full paths
    }
}
\details{
    Creates a lifted hg38 version of the bed file in the current
    directory, with "hg38" interpolated into the filename.
}
\value{
   none
}
\author{
    Paul Shannon
}
\keyword{programming}
